Bugfix in gaussian_kernel

+++ src/stk/filters/gaussian_filter.h
@@ -15,12 +15,12 @@ namespace stk
  * \return A Gaussian filter kernel with radius `r` and size `2r + 1`.
  */
 template <typename T>
-FilterKernel<T> gaussian_kernel(const T sigma, const int spacing)
+FilterKernel<T> gaussian_kernel(const T sigma, const float spacing)
 {
     int r = (int) std::ceil(3 * sigma / spacing); // filter radius
     std::vector<T> kernel (2 * r + 1);
 
-    const T k = -1.0 / (2.0 * sigma * sigma);
+    const T k = -1.0f / (2.0f * sigma * sigma);
     T sum = 0.0;
     for (int i = 0; i < 2*r + 1; ++i) {
         kernel[i] = std::exp(k * (i - r) * (i - r));
